+++ common/tests/com.liferay.ide.ui.tests/src/com/liferay/ide/ui/tests/swtbot/page/AbstractPageObject.java
@@ -47,4 +47,5 @@ public abstract class AbstractPageObject<T extends SWTBot>
     {
         bot.sleep( millis );
     }
+
 }
