+++ common/tests/com.liferay.ide.ui.tests/src/com/liferay/ide/ui/tests/swtbot/page/CancelPageObject.java
@@ -43,4 +43,5 @@ public class CancelPageObject<T extends SWTBot> extends ClosingButtonPageObject<
     {
         return bot.button( text );
     }
+
 }
