+++ common/tests/com.liferay.ide.ui.tests/src/com/liferay/ide/ui/tests/swtbot/page/ClosingButtonPageObject.java
@@ -45,4 +45,5 @@ public abstract class ClosingButtonPageObject<T extends SWTBot> extends ShellPag
 
         button.click();
     }
+
 }
