+++ common/tests/com.liferay.ide.ui.tests/src/com/liferay/ide/ui/tests/swtbot/page/TextPageObject.java
@@ -34,14 +34,11 @@ public class TextPageObject<T extends SWTBot> extends AbstractWidgetPageObject<S
     {
         AbstractSWTBot<? extends Widget> widget = getWidget();
 
-        if( widget instanceof SWTBotText )
-        {
-            SWTBotText swtBotText = (SWTBotText) widget;
+        SWTBotText swtBotText = (SWTBotText) widget;
 
-            bot.waitUntil( new WidgetEnabledCondition( swtBotText, true ) );
+        bot.waitUntil( new WidgetEnabledCondition( swtBotText, true ) );
 
-            swtBotText.setText( text );
-        }
+        swtBotText.setText( text );
     }
 
     @Override
