+++ tools/tests/com.liferay.ide.project.ui.tests/src/com/liferay/ide/project/ui/tests/page/ProjectTreePageObject.java
@@ -13,28 +13,31 @@
  *
  *******************************************************************************/
 
-package com.liferay.ide.ui.tests;
+package com.liferay.ide.project.ui.tests.page;
+
+import org.eclipse.swtbot.swt.finder.SWTBot;
+
+import com.liferay.ide.ui.tests.swtbot.page.TreeItemPageObject;
 
 /**
- * @author Terry Jia
- * @author Ashley Yuan
- * @author Ying Xu
+ * @author Li Lu
  */
-public interface UIBase
+public class ProjectTreePageObject<T extends SWTBot> extends TreeItemPageObject<T> implements ProjectBuildAction
 {
 
-    public final String BUTTON_BACK = "< Back";
-    public final String BUTTON_BROWSE = "Browse...";
-    public final String BUTTON_CANCEL = "Cancel";
-    public final String BUTTON_DELETE = "Delete";
-    public final String BUTTON_FINISH = "Finish";
-    public final String BUTTON_HELP = "Help";
-    public final String BUTTON_NEXT = "Next >";
-    public final String BUTTON_OK = "OK";
+    DeleteProjectDialogPageObject<SWTBot> deleteDialog;
 
-    public final String CHECKBOX_INCLUDE_SAMPLE_CODE = "Include sample code";
+    public ProjectTreePageObject( SWTBot bot, String nodeText )
+    {
+        super( bot, nodeText );
 
-    public final String VIEW_PACKAGE_EXPLORER = "Package Explorer";
-    public final String VIEW_WELCOME = "Welcome";
+        deleteDialog = new DeleteProjectDialogPageObject<SWTBot>( bot );
+    }
 
+    public void deleteProject()
+    {
+        doAction( BUTTON_DELETE );
+
+        deleteDialog.confirmDeleteFromDisk();
+    }
 }
